hook up trip gen modal split and vehicle occupancy table to census tract modalSplits
