+++ frontend/app/components/transportation/trip-generation-tables/modal-splits-vehicle-occupancy.js
@@ -1,4 +1,13 @@
 import Component from '@ember/component';
 
 export default class TransportationTripGenerationTablesModalSplitsVehicleOccupancyComponent extends Component {
+  
+  modalSplitVariablesSubset = [
+    'trans_auto_total',
+    'trans_taxi',
+    'trans_public_bus',
+    'trans_public_subway',
+    'trans_walk',
+  ]
+
 }
