+++ frontend/app/templates/components/transportation/analysis-steps.hbs
@@ -13,8 +13,6 @@
       </div>
     </div>
   {{/link-to}}
-  {{log "in analysis steps"}}
-  {{log project.id}}
   {{#link-to
     "project.show.transportation.existing-conditions"
     project.id
