+++ frontend/app/templates/project/show/transportation/existing-conditions/trip-generation/table.hbs
@@ -1,7 +1,31 @@
 <div class="row">
   <div class="sixteen wide column">
+    <div class="ui top clearing segment">
+      <div class="ui secondary menu">
+        <a
+          class="item {{unless this.isRJTW "active"}}"
+          data-test-censustracts-table-isrjtw="false"
+          {{action (mut this.isRJTW) false}}
+        >
+          Residential
+        </a>
+        <a
+          class="item {{if this.isRJTW "active"}}"
+          data-test-censustracts-table-isrjtw="true"
+          {{action (mut this.isRJTW) true}}
+        >
+          Offices
+        </a>
+      </div>
+    </div>
+  </div>
+</div>
+<div class="row">
+  <div class="sixteen wide column">
     <Transportation::TripGenerationTables::ModalSplitsVehicleOccupancy
-      @analysis={{this.model.transportationAnalysis}}
+      @selectedCensusTractData={{await this.selectedCensusTractData}}
+      @modeLookup={{this.modeLookup}}
+      @commuterModes={{this.commuterModes}}
     />
   </div>
 </div>
